move filemappeddocument to json
