+++ localstack/utils/common.py
@@ -5,7 +5,6 @@ import glob
 import hashlib
 import inspect
 import io
-import json
 import logging
 import os
 import re
@@ -76,6 +75,7 @@ from localstack.utils.http import (  # noqa
 # TODO: remove imports from here (need to update any client code that imports these from utils.common)
 from localstack.utils.json import (  # noqa
     CustomEncoder,
+    FileMappedDocument,
     JsonObject,
     assign_to_path,
     canonical_json,
@@ -232,47 +232,6 @@ PEM_KEY_END_REGEX = r"-----END(.*)PRIVATE KEY-----"
 
 # user of the currently running process
 CACHED_USER = None
-
-
-class FileMappedDocument(dict):
-    """A dictionary that is mapped to a json document on disk.
-
-    When the document is created, an attempt is made to load existing contents from disk. To load changes from
-    concurrent writes, run load(). To save and overwrite the current document on disk, run save().
-    """
-
-    path: Union[str, os.PathLike]
-
-    def __init__(self, path: Union[str, os.PathLike], mode=0o664):
-        super().__init__()
-        self.path = path
-        self.mode = mode
-        self.load()
-
-    def load(self):
-        if not os.path.exists(self.path):
-            return
-
-        if os.path.isdir(self.path):
-            raise IsADirectoryError
-
-        with open(self.path, "r") as fd:
-            self.update(json.load(fd))
-
-    def save(self):
-        if os.path.isdir(self.path):
-            raise IsADirectoryError
-
-        if not os.path.exists(self.path):
-            mkdir(os.path.dirname(self.path))
-
-        def opener(path, flags):
-            _fd = os.open(path, flags, self.mode)
-            os.chmod(path, mode=self.mode, follow_symlinks=True)
-            return _fd
-
-        with open(self.path, "w", opener=opener) as fd:
-            json.dump(self, fd)
 
 
 class PortNotAvailableException(Exception):
